Split out ncbitax into a separate python lib
