+++ requirements-conda.txt
@@ -36,6 +36,7 @@ bedtools=2.27.1
 biopython=1.72
 future==0.16.0
 matplotlib=1.5.3
+ncbitax=0.1.2
 pysam=0.15.0
 pybedtools=0.7.10
 PyYAML=3.12
